+++ src/main/java/io/github/forgestove/create_cyber_goggles/mixin/provider/BeltBlockEntityMixin.java
@@ -41,7 +41,7 @@ public abstract class BeltBlockEntityMixin extends KineticBlockEntity
 		var add = super.addToGoggleTooltip(tooltip, isPlayerSneaking);
 		if (!CCG.CONFIG.goggles.enhancedInfo || getSpeed() == 0) return add;
 		var controllerBE = self().getControllerBE();
-		if (controllerBE != null) TooltipUtil.beltThroughput(tooltip, ((BeltBlockEntityMixin) (Object) controllerBE).ccg$rate);
+		if (controllerBE != null) GoggleTooltipUtil.beltThroughput(tooltip, ((BeltBlockEntityMixin) (Object) controllerBE).ccg$rate);
 		return add;
 	}
 	@Override